First version of step test simulation
Note the following: our current control model isn't like FOPDT very much - we have infinite gain and a 1-to-1 actuator response with a defined rate. The models we've been using aren't a good fit because they're intended for a dynamic system. I think we need to change our control strategy up to be more in line with project guidelines and requirements.